+++ paper/paper.bib
@@ -1,21 +1,71 @@
-@Book{h5pybook,
-        keywords = {python, hdf5},
-        year = {2013},
-        publisher = {O'Reilly},
-        title = {Python and HDF5},
-        author = {Andrew Collette}
+@article{van_der_walt_numpy_2011,
+	title = {The {NumPy} array: a structure for efficient numerical computation},
+	volume = {13},
+	number = {2},
+	journal = {Computing in Science \& Engineering},
+	author = {Van Der Walt, Stefan and Colbert, S Chris and Varoquaux, Gael},
+	year = {2011},
+	pages = {22--30}
 }
 
-@ONLINE{hdf5,
-    author = {{The HDF Group}},
-    title = "{Hierarchical Data Format, version 5}",
-    year = {1997-2017},
-    note = {https://www.hdfgroup.org/HDF5/}
+@book{collette_python_2013,
+	title = {Python and {HDF}5: {Unlocking} {Scientific} {Data}},
+	isbn = {978-1-4919-4501-8},
+	shorttitle = {Python and {HDF}5},
+	abstract = {Gain hands-on experience with HDF5 for storing scientific data in Python. This practical guide quickly gets you up to speed on the details, best practices, and pitfalls of using HDF5 to archive and share numerical datasets ranging in size from gigabytes to terabytes.Through real-world examples and practical exercises, you’ll explore topics such as scientific datasets, hierarchically organized groups, user-defined metadata, and interoperable files. Examples are applicable for users of both Python 2 and Python 3. If you’re familiar with the basics of Python data analysis, this is an ideal introduction to HDF5.Get set up with HDF5 tools and create your first HDF5 fileWork with datasets by learning the HDF5 Dataset objectUnderstand advanced features like dataset chunking and compressionLearn how to work with HDF5’s hierarchical structure, using groupsCreate self-describing files by adding metadata with HDF5 attributesTake advantage of HDF5’s type system to create interoperable filesExpress relationships among data with references, named types, and dimension scalesDiscover how Python mechanisms for writing parallel code interact with HDF5},
+	language = {en},
+	publisher = {O'Reilly Media, Inc.},
+	author = {Collette, Andrew},
+	month = oct,
+	year = {2013},
+	note = {Google-Books-ID: lRCMAQAAQBAJ},
+	keywords = {Computers / Databases / General, Computers / General, Computers / Programming / General, Computers / Programming Languages / Python}
 }
 
-@Misc{pytables,
-  author =    {Francesc Alted and Ivan Vilata and others},
-  title =     {{PyTables}: Hierarchical Datasets in {Python}},
-  year =      {2002--},
-  url = "http://www.pytables.org/"
+@misc{lexy_munroe_dont_2015,
+	type = {blog},
+	title = {Don’t use pickle — use {Camel} / fuzzy notepad},
+	url = {https://eev.ee/release/2015/10/15/dont-use-pickle-use-camel/},
+	urldate = {2018-02-12},
+	journal = {fuzzy notebook},
+	author = {{Lexy Munroe}},
+	month = oct,
+	year = {2015},
+	file = {Don’t use pickle — use Camel / fuzzy notepad:/home/aragilar/.mozilla/firefox/fypobks4.default/zotero/storage/J5E7E3CC/dont-use-pickle-use-camel.html:text/html}
+}
+
+@misc{alex_gaynor_pickles_2014,
+	title = {Pickles are for {Delis}, not {Software}},
[…]
