### 2018-02-13 #575 @aragilar
@arfon I've updated and expanded the `paper.md`, is there any additional changes I should make?
