### 2018-02-14 #575 @arfon
> @arfon I've updated and expanded the paper.md, is there any additional changes I should make?

Thanks. This is looking pretty good. It looks like one or more of the citations are incorrect (the `???`) in the compiled paper.

Next up is to find a reviewer...
