### 2018-02-14 #575 @arfon
@ygrange @mattpitkin would either of you be willing to review this submission for JOSS?
