+++ paper/paper.bib
@@ -44,7 +44,7 @@
 	keywords = {PyCon2014}
 }
 
-@misc{alted_pytables:_2002,
+@misc{pytables,
 	title = {{PyTables}: {Hierarchical} {Datasets} in {Python}},
 	url = {http://www.pytables.org/},
 	author = {Alted, Francesc and Vilata, Ivan and {others}},
@@ -58,7 +58,7 @@
 	year = {1997}
 }
 
-@misc{eevee_lexy_munroe_camel:_2017,
+@misc{eevee_lexy_munroe_camel,
 	title = {camel: {Python} serialization for adults},
 	copyright = {ISC License (ISCL)},
 	shorttitle = {camel},
