### 2018-02-14 #575 @mattpitkin
> @ygrange @mattpitkin would either of you be willing to review this submission for JOSS?

Yes, I'm happy to take this.
