### 2018-02-14 #581 @mattpitkin
Hi @aragilar - could you include explicit instructions for installation in the README and documentation? E.g., something like:
> install the software using:
>
> ```bash
> pip install h5preserve
> ```
>
> or by cloning [this](https://github.com/h5preserve/h5preserve) repository and running:
>
> ```bash
> python setup.py install --user
> ```

Could you also add a requirements file?
