### 2018-02-14 #581 @mattpitkin
Having downloaded h5preserve using pip, I'm trying the example shown [here](https://h5preserve.readthedocs.io/en/latest/Usage.html), and in both Python 2.7.12 and 3.5.2 I get the following error:

```
In [6]: my_cool_experiment = Experiment(np.array([1,2,3,4,5]), 10)

In [7]: with open("my_data_file.hdf5", new_registry_list(registry)) as f:
   ...:     f["cool experiment"] = my_cool_experiment
   ...:
<ipython-input-7-f3958ebfc497> in <module>()
----> 1 with open("my_data_file.hdf5", new_registry_list(registry)) as f:
      2     f["cool experiment"] = my_cool_experiment

TypeError: file() argument 2 must be string, not RegistryContainer
```
If I edit it to use `with open("my_data_file.hdf5", "w") as f:` I get the following error:

```
In [10]: with open("my_data_file.hdf5", "w") as f:
    ...:     f["cool experiment"] = my_cool_experiment
    ...:
<ipython-input-10-63e87c258186> in <module>()
      1 with open("my_data_file.hdf5", "w") as f:
----> 2     f["cool experiment"] = my_cool_experiment

TypeError: 'file' object does not support item assignment
```

Do you know what the problem could be?
