### 2018-02-15 50bde41
DOC: Enable doctest
