+++ tox.ini
@@ -17,6 +17,12 @@ deps=-rdoc-requirements.txt
 commands=
     sphinx-build -W -b html -d {envtmpdir}/doctrees .  {envtmpdir}/html
 
+[testenv:doctest]
+changedir=docs
+deps=-rdoc-requirements.txt
+commands=
+    sphinx-build -W -b doctest -d {envtmpdir}/doctrees .  {envtmpdir}/doctest
+
 [testenv:flake8]
 deps=flake8
 commands=
