### 2018-02-15 d1e3290
DOC: Expand installation docs as recommended by JOSS review
