### 2018-02-15 fab6216
FIX: Update CIs so they use py3 not py2
