+++ setup.cfg
@@ -3,8 +3,8 @@ universal=1
 
 [flake8]
 max-complexity = 15
-# don't require spaces around math ops
 ignore=E226,N802,N803,N806,E402
+builtins=unicode
 
 [versioneer]
 VCS = git
