+++ .travis.yml
@@ -19,25 +19,38 @@ cache:
   directories:
     - $HOME/.cache/pip
 
-env:
-  - TOXENV=py27
-  - TOXENV=py34
-  #- TOXENV=py35
-  #- TOXENV=py36
-  - TOXENV=flake8
-  - TOXENV=pylint
-  - TOXENV=docs
-  - TOXENV=doctest
-  - TOXENV=check-manifest
-  #- TOXENV=checkreadme
 matrix:
   include:
+    - python: 2.7
+      env:
+      - TOXENV=py27
+    - python: 3.4
+      env:
+      - TOXENV=py34
     - python: 3.5
       env:
       - TOXENV=py35
     - python: 3.6
       env:
       - TOXENV=py36
+    - python: 3.5
+      env:
+      - TOXENV=flake8
+    - python: 3.5
+      env:
+      - TOXENV=pylint
+    - python: 3.5
+      env:
+      - TOXENV=docs
+    - python: 3.5
+      env:
+      - TOXENV=doctest
+    - python: 3.5
+      env:
+      - TOXENV=check-manifest
+    - python: 3.5
+      env:
+      - TOXENV=checkreadme
 
 before_install:
   - pip install codecov # for coverage
