### 2018-02-15 ee23928
ENH: Drop Python 2.7
