+++ setup.cfg
@@ -1,10 +1,6 @@
-[wheel]
-universal=1
-
 [flake8]
 max-complexity = 15
 ignore=E226,N802,N803,N806,E402
-builtins=unicode
 
 [versioneer]
 VCS = git
