+++ tox.ini
@@ -4,7 +4,7 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py27,py34,py35,py36,flake8,pylint,docs,check-manifest,checkreadme
+envlist = py34,py35,py36,flake8,pylint,docs,check-manifest,checkreadme
 
 [testenv]
 commands = py.test --cov={envsitepackagesdir}/h5preserve -s {posargs}
