+++ .travis.yml
@@ -21,9 +21,6 @@ cache:
 
 matrix:
   include:
-    - python: 2.7
-      env:
-      - TOXENV=py27
     - python: 3.4
       env:
       - TOXENV=py34
