### 2018-02-15 cbe7111
DOC: State which types are included by default
