### 2018-02-15 #581 @mattpitkin
Regarding the automated testing, I'm not really familiar with tox, but what does it mean when the output says

```
================================ 94 passed, 43 xfailed in 0.78 seconds ================================
```

Is it ok/expected that 43 `xfail`?
