### 2018-02-15 #581 @mattpitkin
Just a minor typo thing - the links to the repo in the ["Contributing"](https://h5preserve.readthedocs.io/en/latest/Contributing.html) doc page point to github.org rather than github.com, so they don't work.
