### 2018-02-15 e3d2ca1
DOC: Check links and coverage of docs
