+++ tox.ini
@@ -16,6 +16,8 @@ changedir=docs
 deps=-rdoc-requirements.txt
 commands=
     sphinx-build -W -b html -d {envtmpdir}/doctrees .  {envtmpdir}/html
+    sphinx-build -W -b linkcheck -d {envtmpdir}/doctrees .  {envtmpdir}/html
+    sphinx-build -W -b coverage -d {envtmpdir}/doctrees .  {envtmpdir}/html
 
 [testenv:doctest]
 changedir=docs
