### 2018-02-16 #581 @mattpitkin
I just have a couple more question/requests?

Is it generally the case that you would define the `registry.loader` and `registry.dumper` functions outside of the class they're referring to, or could/should you have them as methods of the class just called, e.g., `dump` and `load`? The latter seems to make more sense to me, but maybe I'm thinking of these methods in the wrong way.

Would it be possible to give explicit examples of creating your own `RegistryContainer`, and using the `GroupContainer`, as only the `DatasetContainer` is used in the current example? Also, would you be able to shown an example of versioning in action, where you create two different versions, as I'm not entirely sure what it means at the moment?
