### 2018-02-16 #581 @mattpitkin
A couple of minor typos to fix in the `paper.md` file:
 * change "...data structures produce by some modelling..." to "...data structures produce**d** by some modelling..."
 * change "...design flaws in pickle have brought up..." to "...design flaws in pickle have **been** brought up..."
