### 2018-02-16 #581 @mattpitkin
Can you add the `doi = {10.1109/MCSE.2011.37}` to the `van_der_walt_numpy_2011` reference in the `paper.bib` file?
