### 2018-02-17 985ac6b
DOC: Add releasing instructions
