### 2018-02-17 7042e6b
DOC: Provide an example of versioning
