### 2018-02-17 f471371
DOC: Add examples of group and manual registry creation
