### 2018-02-19 #581 @mattpitkin
@aragilar - thanks, the examples look good to me, so I'm happy to sign things off.

@arfon - I'm happy for this to be accepted.
